fixed block button issue
