+++ src/Screens/Users/utils.tsx
@@ -100,4 +100,19 @@ function unBlockUser(userId: string) {
 	});
 }
 
-export { blockUser, unBlockUser, renderAddress, getFarmingType, getLandSizeString, getFarmingTypeColor };
+function formatUnixTimestamp(unixTimestamp: number) {
+
+	if (!unixTimestamp) {
+		return '';
+	}
+	const timestampInMilliseconds = unixTimestamp * 1000;
+	const date = new Date(timestampInMilliseconds);
+
+	const year = date.getFullYear();
+	const month = String(date.getMonth() + 1).padStart(2, '0');
+	const day = String(date.getDate()).padStart(2, '0');
+
+	return `${day}/${month}/${year}`;
+}
+
+export { blockUser, unBlockUser, renderAddress, getFarmingType, getLandSizeString, getFarmingTypeColor, formatUnixTimestamp };
